HREOS-1474: Corrección de ventana de obtención de financiación
Former-commit-id: 7d20ea6c70d100dc61adb9a4e43419a60cc18ae3

+++ recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/model/Formalizacion.java
@@ -112,7 +112,7 @@ public class Formalizacion implements Serializable, Auditable {
     private String numExpediente;
     
     @Column(name = "FOR_CAPITALCONCEDIDO")
-    private Long capitalConcedido;	
+    private Double capitalConcedido;	
 	
     @Version   
 	private Long version;
@@ -297,11 +297,11 @@ public class Formalizacion implements Serializable, Auditable {
 		this.numExpediente = numExpediente;
 	}
 
-	public Long getCapitalConcedido() {
+	public Double getCapitalConcedido() {
 		return capitalConcedido;
 	}
 
-	public void setCapitalConcedido(Long capitalConcedido) {
+	public void setCapitalConcedido(Double capitalConcedido) {
 		this.capitalConcedido = capitalConcedido;
 	}
